+++ motu-web/src/main/java/fr/cls/atoll/motu/web/bll/config/IBLLQueueServerConfigManager.java
@@ -17,6 +17,6 @@ public interface IBLLQueueServerConfigManager {
      * 
      * @return
      */
-    int getRequestDefaultPriority();
+    short getRequestDefaultPriority();
 
 }
